Moved basetask to tasks.py
